+++ src/Extensions/OpenTelemetryExtensions.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using System.Diagnostics;
 using System.Net.NetworkInformation;
 using System.Reflection;
 using Azure.Monitor.OpenTelemetry.Exporter;
@@ -12,8 +11,6 @@ using Microsoft.Extensions.Azure;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
-using OpenTelemetry;
-using OpenTelemetry.Metrics;
 using OpenTelemetry.Resources;
 using OpenTelemetry.Trace;
 
@@ -52,13 +49,18 @@ public static class OpenTelemetryExtensions
 
         services.AddSingleton<ITelemetryService, TelemetryService>();
 
-        if (!TryEnableAzureMonitor(services))
-        {
-            TryEnableOtlp(services);
-        }
+        EnableAzureMonitor(services);
     }
 
-    private static bool TryEnableAzureMonitor(this IServiceCollection services)
+    public static void ConfigureOpenTelemetryLogger(this ILoggingBuilder builder)
+    {
+        builder.AddOpenTelemetry(logger =>
+        {
+            logger.AddProcessor(new TelemetryLogRecordEraser());
+        });
+    }
+
+    private static void EnableAzureMonitor(this IServiceCollection services)
     {
 #if DEBUG
         services.AddSingleton(sp =>
@@ -104,38 +106,5 @@ public static class OpenTelemetryExtensions
 #endif
                 options.ConnectionString = appInsightsConnectionString;
             });
-
-        return true;
     }
-
-    private static bool TryEnableOtlp(this IServiceCollection services)
-    {
-        // TODO (lmolkova): find a better env var name to avoid double-negation since we want an explicit opt-in
-        // OTEL_SDK_DISABLED is a standard env var, but the default is `false` -
-        // https://opentelemetry.io/docs/specs/otel/configuration/sdk-enviroment-variables/#general-sdk-configuration
-
-        var otelDisabled = Environment.GetEnvironmentVariable("OTEL_SDK_DISABLED");
-        if (string.IsNullOrEmpty(otelDisabled) || !bool.TryParse(otelDisabled, out var isDisabled) || isDisabled)
-        {
-            return false;
-        }
-
-        services.AddOpenTelemetry()
-            .ConfigureResource(r => r.AddService("azmcp", serviceVersion: Assembly.GetExecutingAssembly()?.GetName()?.Version?.ToString()))
-            .WithTracing(
-                b => b.AddSource("*")
-                .AddAspNetCoreInstrumentation()
-                .AddHttpClientInstrumentation(o =>
-                {
-                    // TODO (lmolkova) necessary because of https://github.com/Azure/azure-sdk-for-net/issues/37446
-                    o.FilterHttpRequestMessage = _ => Activity.Current?.Parent?.Source?.Name != "Azure.Core.Http" && Activity.Current?.Source?.Name != "System.Net.Http";
-                }))
-            .WithMetrics(b => b.AddMeter("*")
-                .AddAspNetCoreInstrumentation()
-                .AddHttpClientInstrumentation())
-            .WithLogging()
-            .UseOtlpExporter();
-        return true;
-    }
-
 }